add some trait for cleaning code
